Added ability to password protect cloakified file
Added the ability in cloakifyFactory.py that prompts the user to use a password to protect the cloaked file. This will simply set the seed of the random library to the password the user provided, then shuffle the order the lines are written in. For decoding it sets the seed of the random library to the password the user provides and preforms the same shuffling process so it knows how to reconstruct the original base64 payload